Move NameableEnum to util package
