+++ tichu-cli/src/main/java/net/incongru/tichu/simu/cmd/PostActionCommandFactory.java
@@ -1,15 +1,13 @@
 package net.incongru.tichu.simu.cmd;
 
-import com.google.common.collect.Lists;
 import net.incongru.tichu.model.Game;
 import net.incongru.tichu.model.Play;
 import net.incongru.tichu.simu.Simulation;
+import net.incongru.tichu.simu.util.NameableEnum;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.function.Predicate;
-import java.util.stream.Collectors;
 
 import static java.util.function.Predicate.not;
 
@@ -33,36 +31,8 @@ public interface PostActionCommandFactory {
 
     Simulation.PostActionCommand debugPlayerHand(String playerName);
 
-    static interface NameableEnum {
-        List<String> altNames();
-
-        static <E extends Enum<E> & NameableEnum> E byName(Class<E> enumClass, String name) {
-            for (E p : enumClass.getEnumConstants()) {
-                if (name.equalsIgnoreCase(p.name())) {
-                    return p;
-                }
-                // TODO oof -- let's improve this
-                if (Lists.transform(p.altNames(), String::toLowerCase).contains(name.toLowerCase())) {
-                    return p;
-                }
-            }
-            // throw new LineParserException
-            throw new IllegalArgumentException(name + " is not a valid " + enumClass.getSimpleName() + ". Known values are: " + allNamesOf(enumClass));
-        }
-
-        static <E extends Enum<E> & NameableEnum> String allNamesOf(Class<E> enumClass) {
-            final List<String> allNames = new ArrayList<>();
-            for (E p : enumClass.getEnumConstants()) {
-                allNames.add(p.name().toLowerCase());
-                // TODO oof -- let's improve this
-                allNames.addAll(Lists.transform(p.altNames(), String::toLowerCase));
-            }
-            return allNames.stream().collect(Collectors.joining(", "));
-        }
-    }
-
     // TODO this should go elsewhere
-    static enum TemporaryPlayNamesEnum implements NameableEnum {
+    enum TemporaryPlayNamesEnum implements NameableEnum {
         Single, Pair, Triplet,
         FullHouse("full house"),
         Straight("street"),
